refactor(frontend): 🚚 create message feature directory
