+++ app/frontend/src/features/chat/hooks/index.ts
@@ -1,5 +1,3 @@
 import { useChat } from "./use-chat";
-import { usePostMessageMutation } from "./use-post-message";
-import { useSaveMessage } from "./use-save-message";
 
-export { useChat, usePostMessageMutation, useSaveMessage };
+export { useChat };
