+++ app/frontend/src/features/chat/hooks/use-chat.tsx
@@ -1,14 +1,11 @@
-import {
-	messagesByChatIdQueryOptions,
-	userChatsQueryOptions,
-} from "@/features/chat/queries";
+import { userChatsQueryOptions } from "@/features/chat/queries";
+import { messagesByChatIdQueryOptions } from "@/features/message/queries";
 import { useSocket } from "@/hooks";
 import { db } from "@/lib/db";
 import { wsEventDataSchema } from "@shared/types";
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { useCallback, useEffect } from "react";
-import { useSaveMessage } from "./use-save-message";
-import { useUpdateMessage } from "./use-update-message";
+import { useSaveMessage, useUpdateMessage } from "@/features/message/hooks";
 
 export function useChat(chatId: string) {
 	const { addEventListener, removeEventListener, sendMessage } = useSocket();
