+++ app/frontend/src/features/chat/queries.ts
@@ -1,7 +1,5 @@
 import api from "@/lib/api";
-import { db } from "@/lib/db";
 import type { Chat } from "@server/db/chats";
-import type { Message } from "@server/db/messages";
 import { queryOptions } from "@tanstack/react-query";
 
 export const userChatsQueryOptions = queryOptions<Chat[]>({
@@ -13,15 +11,3 @@ export const userChatsQueryOptions = queryOptions<Chat[]>({
 		return (await response.json()).data;
 	},
 });
-
-export const messagesByChatIdQueryOptions = (chatId: string) =>
-	queryOptions<Message[]>({
-		queryKey: ["db/messages-by-chat", chatId],
-		initialData: [],
-		queryFn: async () => {
-			return await db.messages
-				.where("chatId")
-				.equals(chatId)
-				.sortBy("createdAt");
-		},
-	});
