+++ app/frontend/src/features/message/hooks/mutations/use-update-message.tsx
@@ -1,4 +1,4 @@
-import { messagesByChatIdQueryOptions } from "@/features/chat/queries";
+import { messagesByChatIdQueryOptions } from "@/features/message/queries";
 import { db } from "@/lib/db";
 import type { Message } from "@server/db/messages";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
